Insurance_Fraud_Detection_Graph_UI/pyproject.toml
[project]
name = "insurance-fraud-detection-graph-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "plotly>=6.4.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
